requirements.txt
selenium
webdriver-manager
pandas
openpyxl
requests
python-dotenv
